requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
langchain
langchain-openai
langchain-qdrant
qdrant-client
